Revert "Use int 64 to store sample counter value"
This reverts commit 5ff2dd4d63d1275f63e1fd14388d8844c22e1e1a.

+++ hal/properties_vaunt.c
@@ -708,8 +708,8 @@ static int set_trig_time_disable(bool tx, const char *chan, uint32_t val) {
     }                                                                                 \
                                                                                       \
     static int hdlr_rx_##ch##_trigger_sample_counter(const char *data, char *ret) {   \
-        uint64_t val1;  \
-        uint64_t val2;  \
+        uint32_t val1;  \
+        uint32_t val2;  \
         uint64_t val;   \
         read_hps_reg("rxa13", &val2);   \
         read_hps_reg("rxa14", &val1);   \
@@ -1458,8 +1458,8 @@ int check_rf_pll(int ch, int uart_fd) {
     }                                                                           \
                                                                                 \
     static int hdlr_tx_##ch##_trigger_sample_counter(const char *data, char *ret) {   \
-        uint64_t val1;   \
-        uint64_t val2;   \
+        uint32_t val1;   \
+        uint32_t val2;   \
         uint64_t val;   \
         read_hps_reg("txa17", &val2);   \
         read_hps_reg("txa18", &val1);   \
